Peank and Valley
In an array of integers, a "peak" is an element which is greater than or equal to adjacent ingers and a valley is an element which is less than or equal to the adjacent integers. For example: 5,8,6,2,3,4,6 has peaks: 8, 6 and valleys 5,2. Given array of integers , sort the array into an alternating sequence of peaks and valleys:
Input: [5,3,1,2,3] 
output: [5,1,3,2,3]